implement first task 3rd module
