+++ KramDeliveryFood_v3/KramDeliveryFood_v3.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31410.357
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "KramDeliveryFood_v3", "KramDeliveryFood_v3\KramDeliveryFood_v3.csproj", "{345765D0-2256-403A-994D-43B676F37917}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{345765D0-2256-403A-994D-43B676F37917}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{345765D0-2256-403A-994D-43B676F37917}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{345765D0-2256-403A-994D-43B676F37917}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{345765D0-2256-403A-994D-43B676F37917}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {18442BB0-A4CD-4204-98CA-FEAECF3810F5}
+	EndGlobalSection
+EndGlobal
