+++ KramDeliveryFood_v3/KramDeliveryFood_v3/Models/Categories.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace KramDeliverFoodCompleted.Models
+namespace KramDeliveryFood_v3.Models
 {
     public class Categories
     {
