+++ KramDeliveryFood_v3/KramDeliveryFood_v3/Models/Delivers.cs
@@ -1,0 +1,9 @@
+﻿namespace KramDeliveryFood_v3.Models
+{
+    public enum Delivers
+    {
+        NovaPoshta,
+        InTime,
+        UkrPoshta
+    }
+}
