+++ KramDeliveryFood_v3/KramDeliveryFood_v3/Models/OrderStatus.cs
@@ -1,0 +1,11 @@
+﻿namespace KramDeliveryFood_v3.Models
+{
+    public enum OrderStatus
+    {
+        New,
+        Processing,
+        Sent,
+        Delivered,
+        Canceled
+    }
+}
