+++ KramDeliveryFood_v3/KramDeliveryFood_v3/Models/Orders.cs
@@ -1,12 +1,13 @@
 ﻿using System.Collections.Generic;
 
-namespace KramDeliverFoodCompleted.Models
+namespace KramDeliveryFood_v3.Models
 {
     public class Orders
     {
         public List<Products> OrderProducts { get; set; }
         public string PhoneNumber { get; set; }
         public string Address { get; set; }
+        public OrderStatus OrderStatus { get; set; }
 
         public Orders()
         {
