+++ KramDeliveryFood_v3/KramDeliveryFood_v3/Models/Tariffs.cs
@@ -1,0 +1,8 @@
+﻿namespace KramDeliveryFood_v3.Models
+{
+    public enum Tariffs
+    {
+        Fixed,
+        ByPostService
+    }
+}
