+++ KramDeliveryFood_v3/KramDeliveryFood_v3/Program.cs
@@ -1,12 +1,12 @@
 ﻿using System;
 
-namespace KramDeliveryFood
+namespace KramDeliveryFood_v3
 {
-    class Program
+    public class Program
     {
         static void Main(string[] args)
         {
-            Console.WriteLine("Hello World!");
+            
         }
     }
 }
