--- OnlineMarket/KramDeliverFoodCompleted/Models/OrderStatus.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace KramDeliverFoodCompleted.Models
-{
-    public enum OrderStatus
-    {
-        New,
-        Processing,
-        Sent,
-        Delivered,
-        Canceled
-    }
-}
