--- V3_KramDeliveryFood/Data/StoreContext.cs
@@ -1,28 +1,0 @@
-﻿using KramDeliverFoodCompleted.Interfaces;
-using System;
-using System.Collections.Generic;
-
-namespace KramDeliverFoodCompleted.Models
-{
-    public class StoreContext : IData
-    {
-        public IList<Product> BaseProducts { get; set; }
-        public IList<Order> Orders { get; set; }
-        public Order Order { get; set; }
-        public string PhoneNumber { get; set; }
-        public string Address { get; set; }
-        public IList<Product> OrderProducts { get; set; }
-        private readonly ISerializerService _serializerService;
-
-        public StoreContext(ISerializerService serializerService)
-        {
-            BaseProducts = new List<Product>();
-            _serializerService = serializerService;
-        }
-
-        public void InitProducts()
-        {
-            BaseProducts = _serializerService.DoDeserialization<Product>();
-        }
-    }
-}
