--- V3_KramDeliveryFood/Interaction/Reader.cs
@@ -1,45 +1,0 @@
-﻿using System;
-
-namespace KramDeliverFoodCompleted.Interaction
-{
-    public class Reader
-    {
-        public virtual int MakeInput()
-        {
-            var result = 0;
-
-            while (true)
-            {
-                var input = Console.ReadLine();
-
-                if (int.TryParse(input, out result))
-                {
-                    if (input.Length != 1)
-                    { 
-                        Messenger.RepeatInput();
-                        continue;
-                    }
-
-                    break;
-                }
-            }
-
-            return result;
-        }
-
-        public static bool BuyMoreProducts()
-        {
-            BuyerMessenger.ShowBuyMoreMessage();
-            var input = Console.ReadLine();
-
-            while (input != "y" && input != "n")
-            {
-                Messenger.RepeatInput();
-
-                input = Console.ReadLine();
-            }
-
-            return input == "y";
-        }
-    }
-}
