--- V3_KramDeliveryFood/Interfaces/IData.cs
@@ -1,17 +1,0 @@
-﻿using KramDeliverFoodCompleted.Models;
-using System.Collections.Generic;
-
-namespace KramDeliverFoodCompleted.Interfaces
-{
-    public interface IData
-    {
-        IList<Product> BaseProducts { get; set; }
-        IList<Order> Orders { get; set; }
-        Order Order { get; set; }
-        string PhoneNumber { get; set; }
-        string Address { get; set; }
-        IList<Product> OrderProducts { get; set; }
-
-        void InitProducts();
-    }
-}
