--- V3_KramDeliveryFood/Interfaces/ILoggerService.cs
@@ -1,7 +1,0 @@
-﻿namespace KramDeliverFoodCompleted.Interfaces
-{
-    public interface ILoggerService
-    {
-        void AddLog(string input);
-    }
-}
